add comments detailing logic, migrate middleware::custom::tests to frontend::custom
