+++ src/frontend/mod.rs
@@ -14,8 +14,10 @@ use crate::middleware::{
     PodSigner, SELF,
 };
 
+mod custom;
 mod operation;
 mod statement;
+pub use custom::*;
 pub use operation::*;
 pub use statement::*;
